finedu_app/requirements.txt
asgiref==3.8.1
Django>=5.0.6,<5.1
psycopg2-binary>=2.9.9,<2.10
sqlparse==0.5.0
python-dotenv>=1.0.1,<1.1
django-summernote>=0.8.20.0,<0.8.21
Pillow>=10.3.0,<10.4
openpyxl>=3.1.3,<3.2
beautifulsoup4>=4.12.3,<4.13
bleach>=6.1.0,<6.2
certifi>=2024.2.2,<2024.3
charset-normalizer>=3.3.2,<3.4
et-xmlfile>=1.1.0,<1.2
frozendict>=2.4.4,<2.5
html5lib>=1.1,<1.2
idna>=3.7,<3.8
lxml>=5.2.2,<5.3
multitasking>=0.0.11,<0.1
numpy>=1.26.4,<1.27
pandas>=2.2.2,<2.3
peewee>=3.17.5,<3.18
platformdirs>=4.2.2,<4.3
# psycopg2>=2.9.9,<2.10
python-dateutil>=2.9.0.post0,<2.10
pytz>=2024.1,<2024.2
requests>=2.32.3,<2.33
six>=1.16.0,<1.17
soupsieve>=2.5,<2.6
tzdata>=2024.1,<2024.2
urllib3>=2.2.1,<2.3
webencodings>=0.5.1,<0.6
yfinance>=0.2.40,<0.3
folium>=0.16.0,<0.17
django-import-export>=4.0.7,<4.1
# selenium>=4.21.0,<4.22
user-agent>=0.1.10,<0.2
fake-useragent>=1.5.1,<1.6
django-debug-toolbar>=4.4.2,<4.5
django-smart-selects>=1.6.0,<1.7
reportlab>=4.2.0,<4.3
seaborn>=0.13.2,<0.14
matplotlib>=3.9.0,<3.10
plotly>=5.22.0,<5.23